+++ spec/guard/jasmine/util_spec.rb
@@ -14,7 +14,7 @@ describe Guard::Jasmine::Util do
 
       it 'does show that the runner is available' do
         Guard::Jasmine::Formatter.should_receive(:info).with "Waiting for Jasmine test runner at http://localhost:8888/jasmine"
-        util.runner_available?('http://localhost:8888/jasmine')
+        util.runner_available?({ :jasmine_url => 'http://localhost:8888/jasmine', :server_timeout => 15 })
       end
     end
 
@@ -28,7 +28,7 @@ describe Guard::Jasmine::Util do
 
         it 'does show that the runner is not available' do
           Guard::Jasmine::Formatter.should_receive(:error).with "Jasmine test runner isn't available: Connection refused"
-          util.runner_available?('http://localhost:8888/jasmine')
+          util.runner_available?({ :jasmine_url => 'http://localhost:8888/jasmine', :server_timeout => 15 })
         end
       end
 
@@ -40,7 +40,7 @@ describe Guard::Jasmine::Util do
 
         it 'does show that the runner is not available' do
           Guard::Jasmine::Formatter.should_receive(:error).with "Jasmine test runner fails with response code 404"
-          util.runner_available?('http://localhost:8888/jasmine')
+          util.runner_available?({ :jasmine_url => 'http://localhost:8888/jasmine', :server_timeout => 15 })
         end
       end
 
@@ -51,7 +51,7 @@ describe Guard::Jasmine::Util do
 
         it 'does show that the runner is not available' do
           Guard::Jasmine::Formatter.should_receive(:error).with "Timeout waiting for the Jasmine test runner."
-          util.runner_available?('http://localhost:8888/jasmine')
+          util.runner_available?({ :jasmine_url => 'http://localhost:8888/jasmine', :server_timeout => 15 })
         end
 
       end